projects/lapaz-dashboard-20260306130416/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
httpx>=0.28.0
anthropic>=0.42.0
voyageai>=0.3.0
supabase>=2.0.0
